add user indicator as table player
