--- client/src/components/PlayerInfo/playerInfo.module.css
@@ -1,35 +1,0 @@
-.end-game-label{
-    font-size: 1.2rem;
-    padding-left: 0.2rem;
-}
-
-.end-game-label__lose{
-    color: #FEACAC;
-}
-
-.end-game-label__win{
-    color: #69B06C;
-}
-
-.player__name{
-    font-size: 1.4rem;
-}
-
-.player__info{
-    align-self: center;
-    width: max-content;
-}
-
-.player__stat{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-}
-
-.player__bet{
-    display: flex;
-    align-items: center;
-    column-gap: 0.5rem;
-    font-size: 1.2rem;
-}
